refactor: move assumptions and limitation component
